+++ client/components/MapContainer2.jsx
@@ -1,0 +1,25 @@
+import React, { Component } from 'react';
+import { Map, InfoWindow, Marker, GoogleApiWrapper } from 'google-maps-react';
+
+const mapStyles = {
+  width: '100%',
+  height: '100%',
+};
+
+export class MapContainer2 extends Component {
+  state = {};
+  render() {
+    return (
+      <Map
+        google={this.props.google}
+        zoom={8}
+        style={mapStyles}
+        initialCenter={{ lat: 47.444, lng: -122.176 }}
+      />
+    );
+  }
+}
+
+export default GoogleApiWrapper({
+  apiKey: 'AIzaSyDVX3tHxQHbube8HVp4tyZnZASsMoUH694',
+})(MapContainer2);
