+++ client/index.js
@@ -15,7 +15,6 @@ ReactDOM.render(
         <App />
       </Provider>
     </BrowserRouter>
-    ,
   </React.StrictMode>,
   document.getElementById('root')
 );
